pyproject.toml
[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[project]
name = "botlab"
version = "0.1.0"
description = "Minimalny rdzeń projektu PvE cycle controller / simulation framework."
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "OpenAI" }
]
license = { text = "Proprietary" }
dependencies = [
    "PyYAML>=6.0.1"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0"
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I"
]